feat: Separate gauges file
