+++ packages/localization/src/config/translations.json
@@ -3101,7 +3101,6 @@
   "Claim your tokens": "Claim your tokens",
   "After the IFO sales finish, you can claim any IFO tokens that you bought, and any unspent CAKE.": "After the IFO sales finish, you can claim any IFO tokens that you bought, and any unspent CAKE.",
   "The maximum amount of CAKE you can commit to the Public Sale equals the number of your iCAKE, which is based on your veCAKE balance at the snapshot time of each IFO. Lock more CAKE for longer durations to increase the maximum CAKE you can commit to the sale.": "The maximum amount of CAKE you can commit to the Public Sale equals the number of your iCAKE, which is based on your veCAKE balance at the snapshot time of each IFO. Lock more CAKE for longer durations to increase the maximum CAKE you can commit to the sale.",
-  "Go to CAKE Staking": "Go to CAKE Staking",
   "Snapshot at": "Snapshot at",
   "The displayed iCAKE is calculated based on this snapshot time.": "The displayed iCAKE is calculated based on this snapshot time.",
   "Ratio": "Ratio",
